create of Seconds and Milliseconds types

+++ src/main/scala/domain/ConfigChecker.scala
@@ -4,10 +4,10 @@ import utils.Eithers.leftIf
 
 case class BadConfiguration(message: String)
 
-case class Config(name: String, musics: Set[Music], steps: Int)
+case class Config(name: String, musics: Set[Music], steps: Long)
 
 object ConfigChecker:
-  def apply(name: String, musics: Set[Music], steps: Int): Either[BadConfiguration, Config] =
+  def apply(name: String, musics: Set[Music], steps: Long): Either[BadConfiguration, Config] =
     for
       _ <- Either.leftIf(
           steps > 0,
